docs extension
Only works for TomDoc'd CoffeeScript right now.

Needs to work for all languages.

I had to modify handlebars to satisfy 'use strict'

Also had to modify the path stuff in tdoc, which works
in node but not atom. We need some way to install npm packages
or something similar maybe?