Constrain the end of the old screen range to the lastRenderedScreenRow, not maxScreenRow
Constraining the end of the new range max screen row instead of the last rendered row is what allows the buffer to get longer when the last rendered screen row is smaller than the last visible screen row. But constraining the end of the old range to this was causing artifacts. The old range should always be constrained to what is rendered on screen now, not what "could be". What "could be" is the domain of the new range.