Fix editor's lack of cooperation as a flexbox item
Two main changes: First, use absolute position left/right/top/bottom of 0 as a hack to achieve 100% width and height of the flexbox that contains the gutter and scroll view when it is inside an editor that is itself a flexbox item. Now we don't rely on the editor having a definite height because of the absolute positioning. This seemed to confuse the newer flexbox implementation, so I fell back to the old one which seems to work better than last time I tried it for this purpose (it had rendering artifacts before)