Specs can be refreshed and focused
No longer using `_.defaults` to assign `window` globals on the Node context's `global` object, because the window gets reloaded but `global` is persisting between reloads, so we'll only ever store values from the first window. We need to explicitly assign globals on an as-needed basis. Currently, we're only assigning document.