Make `RootView` listen for events before loading packages
This is because RootView listens to some of its own events. It needs
to be first in line to handle its own events because package event
handlers might rely on tree view's event handlers having been run.
This also brings behavior more in line with what we'll experience in
specs.