In requireStylesheet, check loadedThemes for the required stylesheet
I've made requireStylesheet also check the loadedThemes for the
requiredStylesheet.

In the event that we have 2 stylesheets with the same name `editor.css`
and `editor.css` I want the `static/` directory stylesheet to be above
the userTheme stylesheet, so that proper cascading occurs. This commit
also adding class names to the styles.