Set tab index on root-view to 0
Previously if the tree-view was open and was focused and no
editors were open and tab was pressed the document body would
become focused which would not allow events bound to root view
to be triggerable.

This ensures that the root-view always gets focus back when
another view is tabbed away from.