Fix cursor autoscrolling when position doesn't change
When assigning the cursor's position, if we decide we want to autoscroll,
we should do so even if the cursor's position doesn't change.

The previous approach relied on a comparison of the position being
assigned, but this is before it has been clipped. Now we rely on a
boolean return value from the marker position assignment methods, which
tells us if the position has changed. If not, we need to emit our own
event to force autoscrolling.