Make sure file operations are executed after watchers are set.
The fs.watch of node is not ready to receive events until the next run
of the uv loop, so a fs.write immediately after the fs.watch will not
trigger change events.

The only reliable way to test fs.watch is to do file operations in
timers.setTimeout, both process.nextTick and window.setTimeout don't
guarrentee to schedule the job in the next uv loop.