Import variables.less into config.less so it can be compiled alone
We try to compile all stylesheets across the board, but now many of
our stylesheets are imported into a master `atom.less` stylesheet. The
config stylesheet would compile fine when included from `atom.less`,
but alone it fails because it doesn't have the variables. We should
probably only try to pre-compile stylesheets we intend to require
directly… using a manifest maybe?