Add node_submodules to node and package load paths
When integrating atom with a fast-changing npm or package, it can
be helpful to vendor the package as a submodule to avoid the
the need to repeatedly publish it. But we want to treat vendored
packages exactly the same as packages in node_modules in every
other respect.

This commit adds node_submodules to the node require path and the
package search path so that any submodules take precedent over
packages bundled into node_modules via package.json. Once a
package stabilizes, we can publish it, add it to the package.json,
and remove the submodule.