Assign ::activePane on the pane container when panes are focused
This commit switches ::hasFocus from a method to a behavior. The
Focusable defines ::hasFocus by default to be equal to ::focused, but
the intention is that classes that include the mixin can define a custom
::hasFocus behavior to account for their children having focus. This
makes it easy to say a pane "has focus" if the pane itself is focused
or if its active pane item is focused. Then we can subscribe to this
behavior and set the active pane when it changes for any reason.