Add report-licenses task
It logs a report about the licenses of all dependencies to standard out,
including the license type, the source of the information, and the text
of the source if it's not just an entry in the `package.json`.

There are still some licenses that couldn't be automatically harvested
that will need to be covered in the overrides.