Fix spec-task.coffee grunt task to work also on Linux
In its initial state `./script/grunt test` failed to run on Linux
because `spec-task.coffee` was written with only MacOS and Windows in
mind. I've added `else` cases with assumption that they would concern a
linux based system. As I see there are build instructions for FreeBSD
so if somebody could check my changes against that OS I would be
obliged.

Also I have changed a bit the way the package specs are read and how
the concurrency of async.queue is set up because:
1. It was crashing on core specs in original state.
2. There was a possibility of a race condition between queue.push() and
setting the queue.drain: if we manage to push and finish all the tasks
before the drain is set, then the drain function will never be called.