Set permissions on atom binary and make CI options OS independent
On linux specs will fail to execute unless atom binary in `/tmp/` is
marked as executable. I think it's best to set it as a part of build
task as it can be than re-used by other tasks (i.e. install task did
it manually).
Also I've noticed that `--log-file=ci.log` is set only for `win32`.
This smells of Continuous Integration so I thought it would be a good
idea to make this option OS independent so I refactored it to be a
parameter of the test task (if run with `grunt test:ci` then the old
`if win32` logic for ci files will be used). I'm not sure whether the
file should be deleted after every single spec run though (but atom
core crew knows better what CI they're using so I won't argue).