:penguin: .desktop file without capital letter
The .desktop file of Atom was named Atom.desktop but it should be
renamed to atom.desktop because a .desktop should not contain capital
letters as any other .desktop files in /usr/share/applications.

Note that without that, it can be not easy to link the window with the
.desktop file and then a dock/panel could not link a launcher with its
window.