:arrow_up: atom-space-pen-views and enhance deprecated exported views
The atom-space-pen-views library now imports its own version of SpacePen
without a lot of the extensions that we add in the version we export
in Atom. In order to maintain the same functionality in the views we
export, we need to update their prototypes to point at the enhanced
SpacePen views instead.