Updated doc to specify package language
I just added a sentence in the first paragraph: "You can write a package in Javascript or CoffeeScript."

This doesn't seem to be declared anywhere. We can apparently write packages in JS but the word Javascript is nowhere in the doc. index.coffee would presumably imply to readers that packages are written in CoffeeScript, but this should be stated explicitly, and early. (And the JS option will be hugely important to many.)