Introduce ProjectRoot.
Today, Project has three methods that return parallel arrays:

    getPaths()
    getDirectories()
    getRepositories()

Instead, there should be one type, `ProjectRoot`, that encapsulates
all of this information. Then these three methods can be replaced
with one method:

    getProjectRoots()

This commit introduces the `ProjectRoot` type and the `getProjectRoots()`
method. The other three methods are now defined in terms of `getProjectRoots()`.

Note that there are two outstanding methods on `Project` that assume a single
root that seem out of place:

    relativize(fullPath)
    contains(pathToCheck)