Add hook for other implementations of [Git]Repository.
With this change, `Project` registers as a consumer of `"atom-repository-provider"`.
When `atom.project.setPaths()` is called, if a path does not have a `GitRepository`
associated with it, it will see if any of the providers of `"atom-repository-provider"`
can produce an appropriate repository object.

I have a test package with a dummy implementation of `HgRepository` and the following
stanza in its `package.json`:

```
  "providedServices": {
    "atom-repository-provider": {
      "versions": {
        "0.0.1": "createHgRepositoryProvider"
      }
    }
  },
```

I verified that my provider was considered for a path with an Hg repo and that my
`HgRepository` was created and assigned to the `ProjectRoot` that was constructed.