Fix random editor spec now for indented soft-wrap
We previously implemented soft-wrap logic again just for this test.
Since the purpose of this test is just to make sure that mutation occurs
correctly, we now instead construct a fresh editor for comparison so
that changes to soft-wrap logic are reflected.

/cc @as-cii Took care of this, so you don’t need to worry about my
previous comment on your PR. Sorry for the noise.