Measure only when lines stay on screen for > 2 frames
Broadly speaking, measuring characters is a quite cheap task (especially since
we introduced <canvas> and its `measureText` API). However, when scrolling we
repeat this action several times and it's likely for a frame to miss the 16ms
goal: in other words, this means that we cannot afford to always measure newly
built lines.

For this reason, this commit introduces a new algorithm which chooses when
characters should be measured. While scrolling we'll now check if the line is on
screen for at least 2 frames (we can even make this configurable): this will
avoid "measurement thrashing" and, hypothetically, the user won't notice that
we're not measuring a certain line. Benchmarks are quite promising and,
apparently, performance don't degrade at all after these changes.

An alternative implementation could have been to decrease the `didStopScrolling`
timeout and measure right after scrolling stopped: however, that would yield to
a less accurate timing (which is not in sync with `requestAnimationFrame`) and
increases the likelihood of noticing a misplaced graphic element (e.g.
decoration).

Please note that this is still not polished and there are some edge cases which
may need to be addressed. On the other hand it demonstrates what I had in mind
so that we're able to validate this proposal based on real code.

/cc: @nathansobo